Move over to MainWindow component
